backend/pyproject.toml
[project]
name = "fullstack"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "django>=5.2.5",
    "django-cors-headers>=4.7.0",
    "gunicorn>=23.0.0",
    "pyjwt>=2.10.1",
    "python-dotenv>=1.1.1",
]

[dependency-groups]
dev = [
    "mypy>=1.17.1",
    "ruff>=0.12.10",
]
